whatpm/t/webidl/webidl-defs.dat

#data
#errors
#document

#data
        //
#errors
#document

#data
interface abc {
};
#errors
1;non-module definition;s
#document
| interface abc {
| };

#data
module abc {
};
#errors
#document
| module abc {
| 
| 
| };

#data
module abc {
};interface xyz {};
#errors
2;non-module definition;s
#document
| module abc {
| 
| 
| };
| 
| interface xyz {
| };

#data
module abc {
  interface xyz {
  };
};
#errors
#document
| module abc {
| 
| interface xyz {
| };
| 
| };

#data
interface a {};
interface b {};
#errors
1;non-module definition;s
2;non-module definition;s

#data
interface a {};
interface a {};
#errors
2;duplicate qname;m
1;non-module definition;s
2;non-module definition;s

#data
module a {
interface b {};
};
module a {
interface b {};
};
#errors
4;duplicate qname;m
5;duplicate qname;m

#data
module a {};
interface a {};
#errors
2;duplicate qname;m
2;non-module definition;s

#data
module a {};
module _a {};
#errors
2;duplicate qname;m

#data
module a {};
module __a {};
#errors

#data
module a {
interface _b {};
interface _b{}
;
};
#errors
3;duplicate qname;m

#data
[ExceptionConsts, abc] interface x {
  [abc, ExceptionConsts] attribute x b;
  [abc, ExceptionConsts] void m ([abc, ExceptionConsts] in long a);
  [abc, ExceptionConsts] const boolean d = FALSE;
};
[abc, ExceptionConsts=xx] module fd {
  [abc, ExceptionConsts] interface d {
    
  };
  [ExceptionConsts, xxx] exception xx {
    [ExceptionConsts, eee] long b;
  };
};
[abc, ExceptionConsts] const boolean y = TRUE;
[def] typedef a b;
#errors
1;non-module definition;s
1;xattr not applicable;w;ExceptionConsts
1;unknown xattr;u;abc
2;unknown xattr;u;abc
2;xattr not applicable;w;ExceptionConsts
3;unknown xattr;u;abc
3;xattr not applicable;w;ExceptionConsts
3;unknown xattr;u;abc
3;xattr not applicable;w;ExceptionConsts
4;unknown xattr;u;abc
4;xattr not applicable;w;ExceptionConsts
6;unknown xattr;u;abc
7;unknown xattr;u;abc
7;xattr not applicable;w;ExceptionConsts
10;xattr not applicable;w;ExceptionConsts
10;unknown xattr;u;xxx
11;xattr not applicable;w;ExceptionConsts
11;unknown xattr;u;eee
14;xattr not applicable;w;ExceptionConsts
14;unknown xattr;u;abc
14;non-module definition;s
15;unknown xattr;u;def
15;non-module definition;s
#document
| [ExceptionConsts, abc] interface x {
|   [abc, ExceptionConsts] attribute x b;
|   [abc, ExceptionConsts] void m ([abc, ExceptionConsts] in long a);
|   [abc, ExceptionConsts] const boolean d = FALSE;
| };
| 
| [abc, ExceptionConsts=xx] module fd {
| 
| [abc, ExceptionConsts] interface d {
| };
| [ExceptionConsts, xxx] exception xx {
|   [ExceptionConsts, eee] long b;
| };
| 
| };
| 
| [abc, ExceptionConsts] const boolean y = TRUE;
| 
| [def] typedef a b;
